# client/public/openapi.yaml
openapi: 3.0.3
info:
  title: WaCalls API
  version: "1.0.0"
  description: |
    API HTTP do **WaCalls** — chamadas de voz do WhatsApp pelo navegador.

    ## Autenticação
    Não há login/senha. Todas as rotas aceitam o header **`X-Client-Id`** (qualquer
    identificador estável), usado para marcar o "dono" das chamadas e filtrar eventos.

    > ⚠️ A API inicia/controla as chamadas (sinalização), mas o **áudio** exige um cliente
    > WebRTC (navegador) com captura de microfone — não é possível só com chamadas HTTP.
servers:
  - url: /
    description: Mesma origem do painel
...
security:
  - apiKeyAuth: []
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key definida em WACALLS_API_KEY (vazia = sem autenticação)"
  parameters:
...
    Sid:
      name: sid
      in: path
      required: true
      schema: { type: string }
      description: ID da sessão (conta WhatsApp).
...
paths:
...
  /api/sessions/{sid}/recording:
    get:
      tags: [Sessões]
      summary: Estado da gravação de chamadas da conta
      parameters:
        - { $ref: "#/components/parameters/Sid" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { enabled: { type: boolean } }
    put:
      tags: [Sessões]
      summary: Liga/desliga a gravação de chamadas da conta
      description: |
        Quando ligado, TODAS as chamadas desta conta são gravadas. No fim de cada
        chamada o MP3 vira uma **nota privada** na conversa do número no Chatwoot
        (se configurado) e um evento `recording` é enviado no webhook da sessão.
        O áudio também fica em `GET /recordings/{callId}.mp3`.
      parameters:
        - { $ref: "#/components/parameters/Sid" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties: { enabled: { type: boolean, example: true } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { enabled: { type: boolean } }
  /recordings/{id}:
    get:
      tags: [Sessões]
      summary: Baixa o MP3 de uma gravação (rota pública / capability)
      description: "Fora de /api/ (sem API key). O id (callId) é não-enumerável e atua como capability."
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, example: "abc123.mp3" }
      responses:
        "200":
          description: Áudio MP3
          content:
            audio/mpeg: { schema: { type: string, format: binary } }
        "404": { description: Não encontrado }
...